refactor: extract a state hook from the resume input card
Move the saved-doc/upload/paste/profile-import state, effects, and handlers into
a useResumeInput hook, leaving index.tsx as the view that wires the existing
subcomponents (UploadZone, SavedResumeMenu, SaveActions, ProfileUrlInput) to it.
